src/Library/GN.Library.AI/VectorStore/ChromaDb/py/chroma_vectorstore_server/requirements.txt
beautifulsoup4 
sentence-transformers 
chromadb
requests
webdriver-manager
langchain-chroma
langchain-community
sentence-transformers
selenium
pydantic 
SpeechRecognition 
argparse
nats-py
opentelemetry.instrumentation
opentelemetry-instrumentation-fastapi

